+++ sources/Client.cpp
@@ -4,7 +4,7 @@ Client::Client() : BusinessPartner() {
     setStatus(true);
 }
 
-Client::Client(int id, std::string description, int legal_id, Adress adress, int phone, std::string email, bool status) : BusinessPartner(id, description, legal_id, adress, phone, email) {
+Client::Client(int id, std::string description, long long int legal_id, Adress adress, int phone, std::string email, bool status) : BusinessPartner(id, description, legal_id, adress, phone, email) {
     setStatus(status);
 }
